Add files via upload
I first experimented outside of the lesson by creating my own simple mathematical function, which divided numbers. It is interesting to note that numbers that do not divide cleanly (like 5) were rounded down. This is something I have seen happen before, and may be because of how the numbers are stored.
The big mess at the end was my attempt at recursively multiplying. My code is right, for positive numbers. For some reason I could not get it to test right in this software though.. still not sure why. If you plug it into Java it'll work though. Or if you do it by hand.