Add files via upload
This was a lot of fun! The last one (Tautonyms) was a challenge. I'm really proud of the efficiency of the first one. Only one line! Because it will calculate if it is true or false in that line, (return string1[-1:] == string2[-1:]) I can just return that, instead of saying (if true, return true, if false, return false).